Adds card viewing functionality to client
Enables users to view all cards in the game through a new
"View All Cards" button on the home page, navigating to a
dedicated card listing page. Includes necessary API service
methods and data models to fetch and display card information.

Also refactors CreateCardRequestConverter to CardResponseConverter
and moves it to the Contracts project.